Update the copyright.txt
This is the copyright file which is distributed with the
official Blender builds distributed via blender.org project,
with the intent to indicate that the build is done by the
blender foundation.

This change makes the file shorter, avoiding stating agreements
in the file itself, and instead referring to the license.

This matches more closely to how other open source projects
tackle the COPYING file.

The updated content has been reviewed by Ton.

Ref #107295