Transform: Remove 'Resize' fallback from 'Shrink/Fatten'
There are no relevant functional changes.

This commit removes the `Resize` fallback from `Shrink/Fatten`
operation.

This fallback is supposedly executed when the operation is called in a
mode other than mesh editing.

But even the `poll` of the `transform.shrink_fatten` operator restricts
its operation only for edit mode.

Furthermore, the fallback scenario is hardly encountered. To trigger
the fallback, the user would need to explicitly define, in the global
keymap, the `transform.transform` operator with the `Shrink/Fatten`
mode on and call it in a mode other than mesh editing.
(Although `transform.resize` could be used instead).

Therefore, considering the lack of benefits in using this resource, it
is better to remove it and prevent its use. Thus, simplifying the code.